requirements.txt
pytz
django
django-crispy-forms
coverage
Pillow
behave-django
selenium
faker
factory_boy
